+++ ChiTech/ChiMath/SpatialDiscretization/FiniteElement/PiecewiseLinear/pwl_04_mappings.cc
@@ -8,7 +8,7 @@ extern ChiMPI& chi_mpi;
 
 //###################################################################
 /**Provides a mapping of cell's DOF from a DFEM perspective.*/
-int SpatialDiscretization_PWLD::
+int64_t SpatialDiscretization_PWLD::
 MapDOF(const chi_mesh::Cell& cell,
        const unsigned int node,
        const chi_math::UnknownManager& unknown_manager,
@@ -87,7 +87,7 @@ MapDOF(const chi_mesh::Cell& cell,
 
 //###################################################################
 /**Provides a mapping of cell's DOF from a DFEM perspective.*/
-int SpatialDiscretization_PWLD::
+int64_t SpatialDiscretization_PWLD::
 MapDOFLocal(const chi_mesh::Cell& cell,
             const unsigned int node,
             const chi_math::UnknownManager& unknown_manager,
