+++ ChiTech/ChiMath/SpatialDiscretization/FiniteVolume/fv.h
@@ -56,20 +56,20 @@ public:
                             chi_math::UnknownManager& unknown_manager) override;
 
   //04a mappings
-  int MapDOF(const chi_mesh::Cell& cell,
-             unsigned int node,
-             const chi_math::UnknownManager& unknown_manager,
-             unsigned int unknown_id,
-             unsigned int component) const override;
+  int64_t MapDOF(const chi_mesh::Cell& cell,
+                 unsigned int node,
+                 const chi_math::UnknownManager& unknown_manager,
+                 unsigned int unknown_id,
+                 unsigned int component) const override;
 
-  int MapDOFLocal(const chi_mesh::Cell& cell,
+  int64_t MapDOFLocal(const chi_mesh::Cell& cell,
                   unsigned int node,
                   const chi_math::UnknownManager& unknown_manager,
                   unsigned int unknown_id,
                   unsigned int component) const override;
 
-  int MapDOF(const chi_mesh::Cell& cell, unsigned int node) const override;
-  int MapDOFLocal(const chi_mesh::Cell& cell, unsigned int node) const override;
+  int64_t MapDOF(const chi_mesh::Cell& cell, unsigned int node) const override;
+  int64_t MapDOFLocal(const chi_mesh::Cell& cell, unsigned int node) const override;
 
   //04b utils
   size_t GetNumLocalDOFs(chi_math::UnknownManager& unknown_manager) override;
